refactor(groth16): move verifaction from zcash to global crate
