+++ crates/arkworks/Cargo.toml
@@ -23,3 +23,4 @@ rand = { version = "0.7", default-features = false }
 [features]
 default = []
 tests = []
+ink = []
