+++ src/groth16/mod.rs
@@ -1,0 +1,2 @@
+//! Groth16 verifaction
+mod verify;
