+++ src/lib.rs
@@ -54,3 +54,5 @@ pub use arkworks_curve::*;
 pub use matter_labs_curve::*;
 #[cfg(feature = "zcash")]
 pub use zcash_curve::*;
+
+mod groth16;
